refactor de componente admin y player a componente dashboard con respectivas validaciones
