+++ client/src/Rutas/MainRotues.jsx
@@ -5,8 +5,7 @@ import Home from "../pages/Home";
 import Login from "../pages/Login";
 import Register from "../pages/Register";
 import FormUser from "../components/Register/FormUser";
-import Player from "../pages/Player";
-import Admin from "../pages/Admin";
+import Dashboard from "../pages/Dashboard";
 import Shop from "../pages/Shop/Shop";
 import ProtectedRoute from "./ProtectedRoutes";
 import ProductDetail from "../pages/Shop/ProductDetail";
@@ -21,13 +20,13 @@ function MainRoutes() {
       <Route path="/login" element={<Login />} />
       <Route path="check-in" element={<Register />} />
       <Route path="form-user" element={<FormUser />} />
-      <Route path="/products" element={<Shop />}/>
-      <Route path="/products/:id" element={<ProductDetail />}/>
+      <Route path="/products" element={<Shop />} />
+      <Route path="/products/:id" element={<ProductDetail />} />
       <Route
         path="dashboard-player"
         element={
           <ProtectedRoute isAllowed={!!userInfoFirestore}>
-            <Player />
+            <Dashboard />
           </ProtectedRoute>
         }
       />
@@ -37,7 +36,7 @@ function MainRoutes() {
           <ProtectedRoute
             isAllowed={!!userInfoFirestore && userInfoFirestore.isAdmin}
           >
-            <Admin />
+            <Dashboard />
           </ProtectedRoute>
         }
       />
