--- client/src/pages/DashboardWrapper.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function DashboardWrapper() {
-  return <div className="admin__container">DashboardWrapper</div>;
-}
-
-export default DashboardWrapper;
